@@ -17,11 +17,11 @@ dev_dependencies:
     sdk: flutter
   integration_test:
     path: ../../../integration_test
-  pedantic: ^1.10.0-nullsafety.1
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.10.0-56.0.dev <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.12.13+hotfix.5"
