@@ -1,11 +1,11 @@
 name: espresso
 description: Java classes for testing Flutter apps using Espresso.
-version: 0.0.1+9
+version: 0.1.0
 homepage: https://github.com/flutter/plugins/espresso
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
 
 # The following section is specific to Flutter.
 flutter:
