@@ -11,10 +11,10 @@ dependencies:
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
     path: ../
-  http: ^0.12.0
+  http: ^0.13.0
 
 dev_dependencies:
-  pedantic: ^1.8.0
+  pedantic: ^1.10.0
   integration_test:
     path: ../../../integration_test
   flutter_driver:
@@ -24,5 +24,5 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+  sdk: ">=2.12.0-259.9.beta <3.0.0"
   flutter: ">=1.12.13+hotfix.4"
