@@ -2,8 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
 import 'dart:html';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
@@ -11,19 +9,19 @@ import 'package:file_selector_web/src/dom_helper.dart';
 import 'package:file_selector_platform_interface/file_selector_platform_interface.dart';
 
 void main() {
-  group('FileSelectorWeb', () {
+  group('dom_helper', () {
     IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-    DomHelper domHelper;
-    FileUploadInputElement input;
+    late DomHelper domHelper;
+    late FileUploadInputElement input;
 
-    FileList FileListItems(List<File> files) {
+    FileList? createFileList(List<File> files) {
       final dataTransfer = DataTransfer();
-      files.forEach(dataTransfer.items.add);
-      return dataTransfer.files;
+      files.forEach(dataTransfer.items!.add);
+      return dataTransfer.files as FileList?;
     }
 
     void setFilesAndTriggerChange(List<File> files) {
-      input.files = FileListItems(files);
+      input.files = createFileList(files);
       input.dispatchEvent(Event('change'));
     }
 
