@@ -5,9 +5,13 @@ import 'package:flutter/material.dart';
 class GetDirectoryPage extends StatelessWidget {
   void _getDirectoryPath(BuildContext context) async {
     final String confirmButtonText = 'Choose';
-    final String directoryPath = await getDirectoryPath(
+    final String? directoryPath = await getDirectoryPath(
       confirmButtonText: confirmButtonText,
     );
+    if (directoryPath == null) {
+      // Operation was canceled by the user.
+      return;
+    }
     await showDialog(
       context: context,
       builder: (context) => TextDisplay(directoryPath),
