@@ -8,7 +8,11 @@ class OpenTextPage extends StatelessWidget {
       label: 'text',
       extensions: ['txt', 'json'],
     );
-    final XFile file = await openFile(acceptedTypeGroups: [typeGroup]);
+    final XFile? file = await openFile(acceptedTypeGroups: [typeGroup]);
+    if (file == null) {
+      // Operation was canceled by the user.
+      return;
+    }
     final String fileName = file.name;
     final String fileContent = await file.readAsString();
 
