@@ -2,11 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart = 2.9
-
+import 'dart:html';
 import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:mockito/mockito.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:file_selector_platform_interface/file_selector_platform_interface.dart';
 import 'package:file_selector_web/file_selector_web.dart';
@@ -15,29 +13,24 @@ import 'package:file_selector_web/src/dom_helper.dart';
 void main() {
   group('FileSelectorWeb', () {
     IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-    MockDomHelper mockDomHelper;
-    FileSelectorWeb plugin;
-
-    setUp(() {
-      mockDomHelper = MockDomHelper();
-      plugin = FileSelectorWeb(domHelper: mockDomHelper);
-    });
 
     group('openFile', () {
-      final mockFile = createXFile('1001', 'identity.png');
-
       testWidgets('works', (WidgetTester _) async {
+        final mockFile = createXFile('1001', 'identity.png');
+
+        final mockDomHelper = MockDomHelper()
+          ..setFiles([mockFile])
+          ..expectAccept('.jpg,.jpeg,image/png,image/*')
+          ..expectMultiple(false);
+
+        final plugin = FileSelectorWeb(domHelper: mockDomHelper);
+
         final typeGroup = XTypeGroup(
           label: 'images',
           extensions: ['jpg', 'jpeg'],
           mimeTypes: ['image/png'],
           webWildCards: ['image/*'],
         );
-
-        when(mockDomHelper.getFiles(
-          accept: '.jpg,.jpeg,image/png,image/*',
-          multiple: false,
-        )).thenAnswer((_) async => [mockFile]);
 
         final file = await plugin.openFile(acceptedTypeGroups: [typeGroup]);
 
@@ -49,19 +42,21 @@ void main() {
     });
 
     group('openFiles', () {
-      final mockFile1 = createXFile('123456', 'file1.txt');
-      final mockFile2 = createXFile('', 'file2.txt');
-
       testWidgets('works', (WidgetTester _) async {
+        final mockFile1 = createXFile('123456', 'file1.txt');
+        final mockFile2 = createXFile('', 'file2.txt');
+
+        final mockDomHelper = MockDomHelper()
+          ..setFiles([mockFile1, mockFile2])
+          ..expectAccept('.txt')
+          ..expectMultiple(true);
+
+        final plugin = FileSelectorWeb(domHelper: mockDomHelper);
+
         final typeGroup = XTypeGroup(
           label: 'files',
           extensions: ['.txt'],
         );
-
-        when(mockDomHelper.getFiles(
-          accept: '.txt',
-          multiple: true,
-        )).thenAnswer((_) async => [mockFile1, mockFile2]);
 
         final files = await plugin.openFiles(acceptedTypeGroups: [typeGroup]);
 
@@ -78,10 +73,47 @@ void main() {
         expect(await files[1].lastModified(), isNotNull);
       });
     });
+
+    group('getSavePath', () {
+      testWidgets('returns non-null', (WidgetTester _) async {
+        final plugin = FileSelectorWeb();
+        final savePath = plugin.getSavePath();
+        expect(await savePath, isNotNull);
+      });
+    });
   });
 }
 
-class MockDomHelper extends Mock implements DomHelper {}
+class MockDomHelper implements DomHelper {
+  List<XFile> _files = <XFile>[];
+  String _expectedAccept = '';
+  bool _expectedMultiple = false;
+
+  @override
+  Future<List<XFile>> getFiles({
+    String accept = '',
+    bool multiple = false,
+    FileUploadInputElement? input,
+  }) {
+    expect(accept, _expectedAccept,
+        reason: 'Expected "accept" value does not match.');
+    expect(multiple, _expectedMultiple,
+        reason: 'Expected "multiple" value does not match.');
+    return Future.value(_files);
+  }
+
+  void setFiles(List<XFile> files) {
+    _files = files;
+  }
+
+  void expectAccept(String accept) {
+    _expectedAccept = accept;
+  }
+
+  void expectMultiple(bool multiple) {
+    _expectedMultiple = multiple;
+  }
+}
 
 XFile createXFile(String content, String name) {
   final data = Uint8List.fromList(content.codeUnits);
